feat: support for specifying databaseId (#698)
This commit adds support for specifying the databaseId when creating a FirebaseFirestore instance.

This allows users to connect to a specific database within their Firebase project.

The `firestore` function now accepts an optional `databaseId` parameter. If not provided, it defaults to the default database for the project.

Additionally, the `NativeFirebaseFirestoreWrapper` now stores the databaseId and uses it when initializing the Firestore instance.

Co-authored-by: Andrew Reed <andrew_reed@hotmail.com>
Co-authored-by: Nicholas Bransby-Williams <nbransby@gmail.com>